requirements.txt
streamlit
SpeechRecognition
pyttsx3
google-genai